TP-44b update: sql utils
